backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
numpy==2.1.3
Pillow==11.0.0
opencv-python-headless==4.10.0.84
onnxruntime==1.20.1
pydantic==2.10.4
pytest==8.3.4
httpx==0.28.1
